requirements.txt
jax
jaxlib 
distrax
numpyro
jaxoplanet
fitsio
sdss-access
astroquery